% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Batch Gradient Descent}
\usage{
batch(x, y, alpha = 0.004, lambda = 0, max.round = 1000,
  precision = 1e-04)
}
\arguments{
\item{x}{the input matrix of predictors. Each column represents a predictor}

\item{y}{the response vector.}

\item{alpha}{the learning rate - typically this would be set to the optimum value}

\item{lambda}{the regularization parameter - if you want to constrain your parameter estimation to avoid overfitting}

\item{max.round}{the maximum number of iterations - in case of delayed convergence, the function would terminate after max.round iterations}
}
\description{
This function performs parameter estimation for Linear Regression using Batch Gradient Descent
}
\examples{
batch(as.matrix(c(1,2,3,4,5),as.matrix(c(1,2,3,4,5),alpha=0.01)
}
