% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgd.R
\name{BatchGradientDescent}
\alias{BatchGradientDescent}
\title{Batch Gradient Descent}
\usage{
BatchGradientDescent(X, Y, alpha = 4e-04, max.iter = 1000,
  precision = 1e-04)
}
\arguments{
\item{X}{the input matrix of predictors. Each column represents a predictor}

\item{Y}{the response vector.}

\item{alpha}{the learning rate - typically this would be set to the optimum value}

\item{max.iter}{the maximum number of iterations - in case of delayed convergence, the function would terminate after max.iter iterations}

\item{precision}{the precision of the result}
}
\description{
This function performs parameter estimation for Linear Regression using Batch Gradient Descent
}
