% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLoss.R
\name{plotLoss}
\alias{plotLoss}
\alias{plotLoss.rlm}
\alias{plotLoss.default}
\title{plotLoss: Plot loss vs iteration graph}
\usage{
plotLoss(object)

\method{plotLoss}{rlm}(object, ...)

\method{plotLoss}{default}(object, ...)
}
\arguments{
\item{object}{an object of class rlm}

\item{...}{other arguments}
}
\description{
Plot the result of loss function vs number of iterations.
}
\section{Methods (by class)}{
\itemize{
\item \code{rlm}: Plot loss vs iteration of rlm object

\item \code{default}: Plot loss vs iteration
}}

