\name{rcane-package}
\alias{rcane-package}
\alias{rcane}
\docType{package}
\title{
\packageTitle{rcane}
}
\description{
\packageDescription{rcane}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcane}
\packageIndices{rcane}
The package uses 4 different optimization algorithms - batch, stochastic, minibatch and coordinate descent
}
\author{
\packageAuthor{rcane}

Maintainer: \packageMaintainer{rcane}
}
\references{
References would come here - need to add references
}
\keyword{ package }

\seealso{
Need to add see also
}
\examples{
rlm(y ~ x, data=data.frame(x=c(1,2,3), y=c(1,2,3)))
}

