% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgasm.R
\name{rgasm}
\alias{rgasm}
\alias{print.rgasm}
\title{rgasm}
\usage{
rgasm(formula, data, method = "sgd", ...)

\method{print}{rgasm}(object, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{method}{the method to be used. Possible values include "bgd", "sgd", "cgd" and "mini-bgd".}

\item{...}{additional arguments to be passed to the low level regression fitting functions.}

\item{object}{The object to be printed.}
}
\description{
\code{rgasm} is an interface for the optimization functions written in the rcane project.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method for rgasm
}}

\examples{
library(datasets)
rgasm(mpg ~ disp, data = mtcars, alpha = 0.00001)

}
