% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\docType{package}
\name{rlm}
\alias{rlm}
\alias{rlm-package}
\alias{rlm}
\alias{print.rlm}
\title{RCANE}
\usage{
rlm(formula, data, method = "sgd", alpha = 1, max.iter = 1000,
  precision = 1e-04, ...)

\method{print}{rlm}(x, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{method}{the method to be used. Possible values include "bgd", "sgd", "cd" and "mini-bgd".}

\item{alpha}{the learning rate - typically this would be set to the optimum value}

\item{max.iter}{the maximum number of iterations - in case of delayed convergence, the function would terminate after max.iter iterations}

\item{precision}{the precision of the result}

\item{...}{additional arguments to be passed to the low level regression fitting functions.}

\item{x}{The object to be printed.}
}
\description{
RCANE

\code{rlm} is an interface for the optimization functions written in the rcane project.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method for rlm
}}

\examples{
library(datasets)
rlm(mpg ~ disp, data = mtcars, alpha = 0.00001)

}
