% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R, R/fitted.R, R/formula.R, R/print.R,
%   R/resid.R
\name{rlm.summaries}
\alias{rlm.summaries}
\alias{coef.rlm}
\alias{fitted.rlm}
\alias{formula.rlm}
\alias{print.rlm}
\alias{resid.rlm}
\title{Accessing rlm Model Fits}
\usage{
\method{coef}{rlm}(object, ...)

\method{fitted}{rlm}(object, ...)

\method{formula}{rlm}(x, ...)

\method{print}{rlm}(x, ...)

\method{resid}{rlm}(object, ...)
}
\arguments{
\item{object, x}{an object of class rlm}

\item{...}{futher arguments passed to or from other methods.}
}
\description{
All these functions are methods for class \code{"rlm"} objects.
}
