% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{StochasticGradientDescent}
\alias{StochasticGradientDescent}
\title{Stochastic Gradient Descent}
\usage{
StochasticGradientDescent(X, Y, alpha = 1, max.iter = 1000,
  precision = 1e-04)
}
\arguments{
\item{X}{the input matrix of predictors. Each column represents a predictor}

\item{Y}{the response vector.}

\item{alpha}{the learning rate - typically this would be set to the optimum value}

\item{max.iter}{the maximum number of iterations - in case of delayed convergence, the function would terminate after max.iter iterations}

\item{precision}{the minimum difference of Betas between each iterations. If no difference is more than precision, stop the iteration.}
}
\description{
This function performs parameter estimation for Linear Regression using Stochastic Gradient Descent
}
