% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcane.R
\name{rgasm}
\alias{rgasm}
\title{rgasm}
\usage{
rgasm(formula, data, method = "sgd", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{method}{the method to be used. Possible values include "bgd", "sgd", "cgd" and "mini-bgd".}

\item{...}{additional arguments to be passed to the low level regression fitting functions.}
}
\description{
This function is a wrapper for the optimization functions written in the rcane project.
}
\examples{
data <- data.frame(matrix(10:18, 3))
formula = X1~X2+X3
rgasm(formula, data, alpha=0.001)
}
