% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm}
\alias{rlm}
\title{Batch Gradient Descent}
\usage{
rlm(formula, data, method = "batch", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{the data frame containing variables in the model}

\item{method}{the type of gradient descent algorithm to be used}

\item{...}{additional parameters such as learning rate, precision and maximum iterations to be passed to the functin}
}
\description{
This function performs heuristic parameter estimation for Linear Regression using different
gradient descent algorithms
}
