% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm}
\alias{rlm}
\alias{print.rlm}
\title{rlm}
\usage{
rlm(formula, data, method = "sgd", ...)

\method{print}{rlm}(x, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{method}{the method to be used. Possible values include "bgd", "sgd", "cgd" and "mini-bgd".}

\item{...}{additional arguments to be passed to the low level regression fitting functions.}

\item{object}{The object to be printed.}
}
\description{
\code{rlm} is an interface for the optimization functions written in the rcane project.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method for rlm
}}

\examples{
library(datasets)
rlm(mpg ~ disp, data = mtcars, alpha = 0.00001)

}
