% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm}
\alias{rlm}
\title{Batch Gradient Descent}
\usage{
rlm(formula, data, method = "batch", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{the data frame containing variables in the model}

\item{method}{the type of gradient descent algorithm to be used}

\item{alpha}{the learning rate at which the parameters converge to optimal values.}

\item{max.round}{the maximum number of iterations - in case of delayed convergence,
the function would terminate after max.round iterations}
}
\description{
This function performs heuristic parameter estimation for Linear Regression using different
gradient descent algorithms
}
\examples{
rlm(y ~ x, data=something, method="batch")
}
