% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_bgd.R
\name{StochasticGradientDescent}
\alias{StochasticGradientDescent}
\title{Stochastic Gradient Descent}
\usage{
StochasticGradientDescent(X, Y, alpha = 1, lambda = 0, max.round = 1000,
  precision = 1e-04)
}
\arguments{
\item{X}{the input matrix of predictors. Each column represents a predictor}

\item{Y}{the response vector.}

\item{alpha}{the learning rate - typically this would be set to the optimum value}

\item{lambda}{the regularization parameter - if you want to constrain your parameter estimation to avoid overfitting}

\item{max.round}{the maximum number of iterations - in case of delayed convergence, the function would terminate after max.round iterations}
}
\description{
This function performs parameter estimation for Linear Regression using Stochastic Gradient Descent
}
\examples{
StochasticGradientDescent(as.matrix(c(1,2,3,4,5), as.matrix(c(1,2,3,4,5), alpha=0.01)
}
